requirements.txt
flask
flask-cors
tensorflow==2.9.1
pillow
gunicorn
numpy==1.26.4
tflite-runtime